+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Guide: https://keepachangelog.com/en/1.0.0/
 
 ## [Unreleased]
 
-## [1.0.0-beta.21] - 2022-01-19
+## [1.0.0-beta.22] - 2022-01-19
 
 ### Added